productionize spf tool
