+++ Sources/HTML/test.spf.swift
@@ -1,0 +1,545 @@
+import Glibc
+
+/// this is a doccomment
+enum A:Sendable 
+{
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+}
+
+/// this is a doccomment
+enum B:Sendable 
+{
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+}
+
+/// this is a doccomment
+enum C:Sendable 
+{
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``A``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.A.string)
+    /// """
+    /// ```
+    enum A:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``B``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.B.string)
+    /// """
+    /// ```
+    enum B:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:X
+    {
+
+        static 
+        var string:String { "xxx" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:Y
+    {
+
+        static 
+        var string:String { "yyy" }
+    }
+    /// this is another doccomment
+    /// with many lines, and it 
+    /// mentions ``C``.
+    /// 
+    /// ```swift
+    /// // nested comments and string 
+    /// // interpolations work!
+    /// let string:String = 
+    /// """
+    /// \(Self.C.string)
+    /// """
+    /// ```
+    enum C:Z
+    {
+
+        static 
+        var string:String { "zzz" }
+    }
+}
+protocol X:Sendable 
+{
+    var X:Int { get }
+}
+protocol Y:Sendable 
+{
+    var Y:Int { get }
+}
+protocol Z:Sendable 
+{
+    var Z:Int { get }
+}
+
+extension DOM.Element where Domain == HTML 
+{
+    @inlinable public 
+    var i:Int 
+    {
+        0
+    }
+    @inlinable public 
+    var j:Int 
+    {
+        1
+    }
+    @inlinable public 
+    var k:Int 
+    {
+        2
+    }
+
+    @inlinable public static 
+    func a(_ string:String, @Attributes attributes:() -> [Attribute] = { [] }) -> Self 
+    {
+        .a(.text(escaping: a.rawValue), attributes: attributes)
+    }
+
+    @inlinable public static 
+    func b(_ string:String, @Attributes attributes:() -> [Attribute] = { [] }) -> Self 
+    {
+        .b(.text(escaping: b.rawValue), attributes: attributes)
+    }
+}
